Major Battle update
pushing more setup for battles. Added in functionality for setuping up battles including payout and any modifiers. Also added a join battle command searches for user and adds their information to the battle including health damage and armor. Need to setup function to run the battle and do the math to calculate the winner. Also need to reset the battle afterwords